post create

+++ backend/myapp/views.py
@@ -114,14 +114,14 @@ class UserLoginView(generics.GenericAPIView):
             return Response(response_data, status=status.HTTP_200_OK)
         return Response(serializer.errors, status=status.HTTP_400_BAD_REQUEST)
     
-    def patch(self, request, *args, **kwargs):
-        # Update existing user
-        user = self.get_object()  # Get the current user
-        serializer = self.get_serializer(user, data=request.data, partial=True)
-        if serializer.is_valid():
-            user = serializer.save()
-            return Response({'status': 'User updated'}, status=status.HTTP_200_OK)
-        return Response(serializer.errors, status=status.HTTP_400_BAD_REQUEST)
+    # def patch(self, request, *args, **kwargs):
+    #     # Update existing user
+    #     user = self.get_object()  # Get the current user
+    #     serializer = self.get_serializer(user, data=request.data, partial=True)
+    #     if serializer.is_valid():
+    #         user = serializer.save()
+    #         return Response({'status': 'User updated'}, status=status.HTTP_200_OK)
+    #     return Response(serializer.errors, status=status.HTTP_400_BAD_REQUEST)
 # logger = logging.getLogger(__name__)
 class IsAuthenticatedOrReadOnlyForUpdateDelete(BasePermission):
     """
@@ -220,11 +220,8 @@ class PostCategoryModelViewSet(viewsets.ModelViewSet):
 class PostModelViewSet(viewsets.ModelViewSet):
     serializer_class = PostSerializer
     permission_classes = [IsAuthenticatedOrReadOnly]
-    queryset = Post.objects.all()
-
 
     def get_queryset(self):
-        # This method is still responsible for querying
         queryset = Post.objects.all()
         category_id = self.request.query_params.get('category_id', None)
         
@@ -241,69 +238,38 @@ class PostModelViewSet(viewsets.ModelViewSet):
         else:
             cache_key = 'post_list'
         
-        # Check if the data is in cache
         cached_data = cache.get(cache_key)
         
         if cached_data is not None:
-            print('Cache hit',cached_data,cache_key)
-            # Create a Response object with the cached data
+            print('Cache hit', cached_data, cache_key)
             response = Response(cached_data)
         else:
-            # Fetch data from the database
             queryset = self.get_queryset()
             data = PostSerializer(queryset, many=True, context={'request': request}).data
-            
-            # Cache the data
-            cache.set(cache_key, data, timeout=60*2)  # Cache for 2 minutes
+            cache.set(cache_key, data, timeout=60*2)
             print('Cache miss')
-            
-            # Create a Response object with the fetched data
             response = Response(data)
         
         return response
     
-    # def retrieve(self, request, *args, **kwargs):
-    #     print("Request:", request)
-    #     print("Arguments:", args)
-    #     print("Keyword Arguments:", kwargs)
-    #     cache_key = f'post_{kwargs["pk"]}'
-    #     data = cache.get(cache_key)
-    #     if not data:
-    #         response = super().retrieve(request, *args, **kwargs)
-    #         data = response.data
-    #         cache.set(cache_key, data, timeout=60 * 15)  # Cache for 15 minutes
-    #     else:
-    #         response = Response(data)
-    #     return response
-    
     def perform_create(self, serializer):
-        # Save the new post
         instance = serializer.save(user=self.request.user)
-        
-        # Invalidate the cache
-        cache.delete('post_list')  # Invalidate the cache for the post list
-        cache.delete(f'post_category_{instance.post_category_id}')  # Invalidate category-specific cache
+        cache.delete('post_list')
+        cache.delete(f'post_category_{instance.post_category_id}')
 
     def perform_destroy(self, instance):
-        # Delete the image file from the filesystem
         if instance.image:
-            if os.path.isfile(instance.image.path):  
-                os.remove(instance.image.path)
-        # Delete the Post instance
-        instance.delete()
-        # Invalidate the cache
-        cache.delete('post_list')  # Invalidate the cache for the post list
-        cache.delete(f'post_category_{instance.post_category_id}')  # Invalidate category-specific cache
-
-    def perform_update(self, serializer):
-        # Check if the image is being updated
-        instance = self.get_object() 
-        if 'image' in self.request.FILES and instance.image:
-            # Delete the old image file from the filesystem
             if os.path.isfile(instance.image.path):
                 os.remove(instance.image.path)
-        # Save the new instance with the updated data
+        instance.delete()
+        cache.delete('post_list')
+        cache.delete(f'post_category_{instance.post_category_id}')
+
+    def perform_update(self, serializer):
+        instance = self.get_object()
+        if 'image' in self.request.FILES and instance.image:
+            if os.path.isfile(instance.image.path):
+                os.remove(instance.image.path)
         serializer.save()
-        # Invalidate the cache
-        cache.delete('post_list')  # Invalidate the cache for the post list
-        cache.delete(f'post_category_{instance.post_category_id}')  # Invalidate category-specific cache
+        cache.delete('post_list')
+        cache.delete(f'post_category_{instance.post_category_id}')
